build(macos): configure and compile toward a native macOS arm64 build
- Add a macos-clang preset pinned to Xcode's clang; apply -march=sandybridge
  only on x86_64 hosts.
- Enable OBJC/OBJCXX at the top level so plume's Metal backend generates.
- Find Python3 in the runtime CMake; macOS ships no bare `python`.
- Add an asm-free Darwin config for the pinned Xenia FFmpeg fork and skip
  its x86 init sources on Apple.
- Extend POSIX CRT shims to macOS; declare environ, resolve the executable
  directory with _NSGetExecutablePath, and use libc++ file_clock::to_sys.
- Ignore .DS_Store and local ISO images.

+++ LostOdysseyRecomp/gpu/shader/resource_scan.h
@@ -296,13 +296,13 @@ inline Result Scan(const fs::path& root, const fs::path& cacheDir,
         for (const auto& file : files) {
             const auto size = fs::file_size(file); totalBytes += size;
             identity << fs::absolute(file).generic_string() << '\t' << size << '\t'
-                     << fs::last_write_time(file).time_since_epoch().count() << '\n';
+                     << static_cast<long long>(fs::last_write_time(file).time_since_epoch().count()) << '\n';
         }
         // FPI changes also invalidate the discovery cache. Like FPD identity,
         // this is a size/mtime fingerprint, not a full game-integrity digest.
         for (const auto& file : indexes)
             identity << fs::absolute(file).generic_string() << '\t' << fs::file_size(file) << '\t'
-                     << fs::last_write_time(file).time_since_epoch().count() << '\n';
+                     << static_cast<long long>(fs::last_write_time(file).time_since_epoch().count()) << '\n';
         const auto source = cacheDir / "source";
         fs::create_directories(source);
         const auto manifest = cacheDir / "resources.manifest";
